Integrate Balancing Strategy for Partitioning with Vertex Weights (#70)
* fix heavy assertions

* some small fixes

* tools for creating weighted instances

* add segment tree to datastructures

* algorithms for bin packing

* test for bin packing

* bin packing initial partitioner

* add tests for resultingMaxBin

* implementation of balancing strategy´

* cmd args and current configs

* some additional statistics

* fix for fixed vertices

* some renaming

* bug fix for prepacking algorithm

wrong parantheses, affecting case with unequal ks

* fix coarseners to handle fixed vertices correctly

* improved output and commentary on open aspects

* fix for some tests

* update old configs

* simplification for first fit algorithm

* more unit tests

* many consts

* integration test for node weights

* update context creation in tests

* add casts

* simplify segment tree implementation

* some renaming

* restructure bin packer

* many small changes from review, formatting

* separate file for i_bin_packer

* docu and more formatting

* Revert "update old configs"

This reverts commit 1852c6b3eef8b8d7dbf5f02b2d7f5ab754440d4c.

* fix integration tests

* statistics for hyperedge weights

* more test fixes

* typo

* test fix

* remove part of algorithm that is not required

* adjust test

* remove TODO

* types in segment tree, namespace

* recursive node removal with included epsilon

* improved node removal

* alternative balance constraint formulation

prepacking implementation and tests

* use factory to create bin packer

* disable fixed vertex imbalance warning in legit cases

* add missing includes

* fix for segment tree

* integration test with smaller instance, use working configs

* command line option to disable heuristic prepacking

* update configs and tests

* minor

* refactor late restart

+++ kahypar/datastructure/binary_heap.h
@@ -60,7 +60,7 @@ class BinaryHeapBase {
     _compare(),
     _next_slot(0),
     _max_size(size + 1) {
-    for (size_t i = 0; i < size; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(size); ++i) {
       _heap[i] = HeapElement(BinaryHeapTraits<Derived>::sentinel());
       _handles[i] = 0;
     }
